Made some client islands for about
